internal/gosym: fork debug/gosym
This is the first CL in a sequence that will allow vulncheck to
see inlined symbols in a binary.

In order to do this without relying on DWARF debugging information, we
need to traverse some data structures in the binary file. debug/gosym
does most of the work, but we need to extend it to find information
about inlined functions.

We start by forking debug/gosym and getting it to work outside of the
standard library.

(Also, make check.bash skip license checks for files in testdata
directories.)

For golang/go#51412.

Change-Id: I21d2911fde59dd8eb8f0e6d82bc28365f147c68d
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/398295
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>