cmd/govulncheck: test the command directly
Add a test that builds govulncheck and compares its output to a golden.

This uses the github.com/google/go-cmdtest package, which makes it easy
to do that.

Move the buildtest package out of vulncheck/internal so cmd/govulncheck
can use it, and fix a bug when GoBuild is passed ".".

Change-Id: I2e216f1980cfe0548e7231531bd68b2996142917
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/399115
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>