vulncheck: return the module list
As a convenience, have vulncheck return the list of modules it found
in the binary or source.

When vulcheck is run on source, this information duplicates information
in the requires graph, although in a simpler form.

When it is run on a binary, the requires graph is nil, so this
information is new.

The list of modules and their versions can be found in other ways in
both cases, but it would require duplicating work that vulncheck
already does.

This list will be used by the govulncheck command to display the
versions currently in use.

Change-Id: I35fe10a456ca7d5265340314a2aba402e648af10
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/399934
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>