cmd/govulncheck: add unaffected information
Modules that contain vulnerabilities that are not called are now printed
at the end of a default output. This information makes it clear to the
user that we didn't miss a vulnerability, and rather are actively
informing them that these vulnerabilities are irrelevant to them.

Change-Id: I162a976eba6554ffbefca77178e27e3332db871b
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/409814
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Julie Qiu <julieqiu@google.com>
Run-TryBot: Julie Qiu <julieqiu@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>