cmd/govulncheck: inline LegacyRun into govulncheck main
This is the second in the series of CLs that attempts to refactor
govulncheck. The changes in this CL are:

- inline the logic of LegacyRun into cmd/govulncheck/main.go
- delete LegacyConfig and related data structures
- move all the accompanying logic to cmd/govulncheck
- expose db caching logic

This effectively results in the following structure.
internal/govulncheck contains the core logic for Source and Binary
analysis (including db caching). cmd/govulncheck contains the logic for
parsing the flags and printing the results.

The follow-up refactoring CLs will mainly focus on renaming of things
and addressing TODOs.

For golang/go#56042

Change-Id: Ib1fc26951420ae6d04e48205765a1019d117946f
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/447855
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>