cmd/govulncheck/integration: adjust expectations for new db vulns
The changes in expectations are due to
https://go-review.git.corp.google.com/c/vulndb/+/468899,
https://go-review.git.corp.google.com/c/vulndb/+/468900,
https://go-review.git.corp.google.com/c/vulndb/+/468555, and
https://go-review.git.corp.google.com/c/vulndb/+/468898.

This CL also changes the granularity of expectations. We now check just
for packages of called/present symbols. It is not realisitic to check
that every symbol is called. Also, checking for every symbol requires
more maintenance that in addition happen more frequently.

Change-Id: I047150126ad6b2cb474e78619cabcf7e197cbbf5
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/469098
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>