internal/govulncheck: return exit code 0 for -json mode when vulns are found
When vulnerabilities are found, we return an exit code 3 if running in
text mode, because we assume that users are checking for whether
vulnerabilities exist in their code. Therefore, the presence of
vulnerabilities should be a failure, regardless of what is in the
output.

This isn't the case for -json mode, since the presence of JSON data
indicates success, so return an exit code of 0. This is consistent with
a decision from CL 432236.

Change-Id: Ie403ea79e40eb80bad590679a27ef9c3f5f8665a
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/474795
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Julie Qiu <julieqiu@google.com>
Reviewed-by: Ian Cottrell <iancottrell@google.com>
Run-TryBot: Julie Qiu <julieqiu@google.com>