internal/scan: add mode flag
At the moment, govulncheck tries to be smart about detecting when
something is a binary, but when it gets this wrong, there is no way for
the user to signal that to the tool.

Add a -mode flag to explicitly distinguish between source and binary
mode. By default, govulncheck always runs source analysis.

Change-Id: I476ac245210d84a4f7e40df7f6896d7b633eb1cb
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/481137
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Julie Qiu <julieqiu@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>