Revert "cmd,internal/scan: remove verbose mode"
This reverts commit ebda8dc8a2d6000060ede47302142dee6a10319a.

Reason for revert: There are a couple of folks who have reported finding the -v flag useful, so revert this CL. We don't think that -v is the right flag for the information surfaced in this mode, but the information currently surfaced in verbose mode does seem to be useful. We plan revisit and post updates to a GitHub issue.

Change-Id: I72dc002cf90bc746230cc3f3fbf2e957e6a7c5de
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/482899
Run-TryBot: Julie Qiu <julieqiu@google.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>