internal/vulncheck/internal/buildinfo: deduplicate extracted symbols
The symbol extraction would put in a list every use of a symbol in a
binary. So if the symbol is used multiple times, it will appear in the
list several times. This CL removes that redundancy.

Updates golang/go#59696

Change-Id: I972ee03d317a5d0b831b36151bac7079b745adf7
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/486155
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>