internal/scan: remove redundant symbols from binary results
In source mode, we try to avoid showing redundant call stacks and
redundant symbols. For instance, if a symbol "s" is reachable only
through other symbols, then "s" is not reported.

This CL attempts a similar approach for the binary mode. Due to the lack
of call stacks in this mode, the following approach is used. No
unexported symbols are reported if some exported symbols are present
in the binary for a vulnerability. Otherwise, all unexported symbols
are reported. The latter case is not expected to happen often.

Fixes golang/go#59696

Change-Id: Ic5db40a3bc4a7bbfd9818338ca82c46662951c31
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/486195
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Julie Qiu <julieqiu@google.com>