internal/client: change error message for invalid db and small refactor
Changes the error message for DBs that do not implement the v1 schema
to remove reference to the legacy schema.

Also changes the API of the newLocalSource function to take in a
a directory, as the directory is now already computed by the calling
function.

Change-Id: Ic6bc86084aeaf4d513e417e7c5c8805a984c740c
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/494404
Reviewed-by: Cherry Mui <cherryyz@google.com>
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>