internal/client: allow flat list of OSV as db
Allows a user of govulncheck to pass in a folder containing OSV JSON
files as a database.

Under the hood, we index the files and create an in-memory index. The
bulk of this logic is in a new type, hybridSource.

This is currently a hidden, experimental feature.

Change-Id: Ic1cc5f02badf2e3a51024eb3e6e9e3d7cf69b576
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/494405
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Julie Qiu <julieqiu@google.com>
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>