internal/gosym: update binary mode version parsing
This changes the behavior of govulncheck when ran on a compiled binary.
Previously, if the binary was built with an invalid version or < v1.18,
gosym.FuncSymName would return an empty string. Now, it only returns an
empty string when the binary has been built with version < 1.18, and
assumes that unparseable versions are newer than 1.20 and returns the
function symbol name for go version 1.20 or above.

Change-Id: Id53b513802a5629dd0b2aea824424251ae1512ed
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/514901
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Run-TryBot: Maceo Thompson <maceothompson@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>