internal/semver: rename the LatestFixedVersion function
The name might suggest that this function computes the most recent
version with a fix for a vulnerability. But in fact, this function
is doing more; it is also considering if the most recent fixed version
is superseded by a re-introduction of a vulnerability. The name of
the function is thus changed accordingly.

Change-Id: I96e58ce128b5f59f21358e82d5a74a388cdd09b3
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/530117
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>