internal/vulncheck: update logic for package level analysis
We are not computing a proper slice of vulnerable imports graph, so we
update the documentation to reflect that. Also, we never don't really
have a need to maintain package entry points, so we get rid of them.

Some related cleaning and renaming is done as well.

This CL is part of a series of changes cleaning up the code a bit.

Change-Id: I3b530c0a74b65933ec0c20f6f86a0a9c11b66929
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/539340
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Maceo Thompson <maceothompson@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>