internal/scan: add scan level to textHandler
This change adds scan level to the text handler, which will allow us to
change messages depending on the amount of information govulncheck has
available to it. For example, it will allow govulncheck to not mention
callstacks if it never ran callstack analysis.

Change-Id: I9dccd5b0445dddfd8c071bf266ab9eecb0bd0d5e
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/545636
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>