all: remove test that runs govulncheck on govulncheck
The only interesting vulnerabilities this test can find are stdlib ones.
We cannot really do much about those. On the other hand, whoever is
developing this module on a go version that is not the latest patch or
top will see a failure. This is annoying so we remove the test.

Change-Id: I3236393354565eb9d682075948330b0612434606
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/589235
Reviewed-by: Maceo Thompson <maceothompson@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>