internal/vulncheck: use module info when looking for symbols
When searching if (a symbol of a) package is vulnerable, we would try to
identify the module from package path. (The module information is needed
because we save vulns per module.) This can cause problems when module
paths are prefixes of each other. In all cases except binary mode, we
know the exact module of a symbol or a package, so we simply use that.

Change-Id: I21c220e485522dda1bc1fe0a9025e73846b6fd6f
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/592135
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Maceo Thompson <maceothompson@google.com>