internal/semver: add SemverToGoTag
Add SemverToGoTag, which reverses GoTagToSemver.
The exception is when converting a go tag to the semver would have
resulted in an invalid semver (such as go RC versions). GoTagToSemver
returns an empty string in that case, and as such is not reversible.

Change-Id: I8d68c2372261b1ec7ec8b72354ac1919ad24d6f6
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/617396
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>